rename App.tsx to index.tsx
