+++ packages/client/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app/App';
+import App from './app';
 import './index.pcss';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
